requirements.txt
aniso8601==7.0.0
asgiref==3.3.4
autopep8==1.5.7
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
Django==3.2.2
django-admin-interface==0.16.3
django-appconf==1.0.4
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-cryptography==1.0
django-encrypted-filefield==0.3.0
django-environ==0.4.5
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-jazzmin==2.4.7
django-markdownfield==0.10.0
djangorestframework==3.12.4
graphene==2.1.8
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
Markdown==3.3.4
packaging==20.9
Pillow==8.2.0
promise==2.3
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pyparsing==2.4.7
python-magic==0.4.22
pytz==2021.1
requests==2.25.1
Rx==1.6.1
shortuuid==1.0.1
singledispatch==3.6.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.5
webencodings==0.5.1
